+++ docs/CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   {py:func}`salespyforce.utils.version.get_major_minor_version` functions now attempt to 
   retrieve the version from the `pyproject.toml` file if it cannot be retrieved via the 
   package metadata.
+- The {py:meth}`salespyforce.Salesforce.Chatter.post_feed_item` and
+  {py:meth}`salespyforce.Salesforce.Chatter.post_comment` methods (and their underlying functions) 
+  now raise the {py:exc}`~salespyforce.errors.exceptions.MissingRequiredDataError` exception rather 
+  than the generic {py:exc}`RuntimeError` exception when required data is not passed appropriately.
 - The documentation for this project has updated to a more modern and intuitive theme,
   layout, and structure. The intent was to be less verbose and instead make the documentation 
   more human-oriented with helpful guides and tutorials.